changed homepga colours to make more thematically appropriate and made ratings integers

+++ products/models.py
@@ -23,7 +23,7 @@ class Product(models.Model):
     description = models.TextField()
     has_sizes = models.BooleanField(default=True, null=True, blank=True)
     price = models.DecimalField(max_digits=6, decimal_places=2)
-    rating = models.DecimalField(max_digits=6, decimal_places=2, null=True, blank=True)
+    rating = models.IntegerField(null=True, blank=True)
     image_url = models.URLField(max_length=1024, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
